+++ sessions/nKsh7ORmHAc8ilTcpPKmvGrVJnrB7z0o.json
@@ -1,1 +1,1 @@
-{"cookie":{"originalMaxAge":null,"expires":null,"httpOnly":true,"path":"/"},"__lastAccess":1647435771593,"loggedUser":{"id":34,"nickname":"aaa11","updatedAt":"2022-03-16T12:56:48.398Z","createdAt":"2022-03-16T12:56:48.398Z"}}
+{"cookie":{"originalMaxAge":null,"expires":null,"httpOnly":true,"path":"/"},"__lastAccess":1647437674283,"loggedUser":{"id":43,"nickname":"dd1","updatedAt":"2022-03-16T13:34:21.763Z","createdAt":"2022-03-16T13:34:21.763Z"}}
